migrate testsuite from nose to pytest
- nose is no longer maintained
- add minimal github actions CI to install pybio-utils and run these tests
- also had to change the order of items in these objects to make tests pass:
  - `expected_transcript_fractions` list
  - `expected_bed_overlaps list`
  - `expected_gtf_entries attributes` string
