+++ setup.cfg
@@ -33,12 +33,13 @@ install_requires =
     matplotlib_venn
     seaborn
 python_requires = >=3.6,<3.7.0a0
-test_suite =
-    nose.collector
-tests_require =
-    nose
 include_package_data = True
 zip_safe = False
+
+[options.extras_require]
+test =
+    pytest
+    pytest-cov
 
 [options.entry_points]
 console_scripts =
@@ -85,9 +86,6 @@ console_scripts =
     select-periodic-offsets = pbio.ribo.pgrms.select_periodic_offsets:main
     # misc
     pickle-stan = pbio.misc.pgrms.pickle_stan:main
-
-# config for test integration
-# [nosetests]
 
 # sphinx integration with setuptools
 # [build_sphinx]
